requirements.txt
# Core dependencies for SyncSpace Algo - Vision-Based Processing
# Document processing
python-docx>=1.1.0
openpyxl>=3.1.0
pandas>=2.0.0
pypdf>=4.0.0

# Image processing for Excel to image conversion
Pillow>=10.0.0

# Web frameworks
Flask>=3.0.0
fastapi>=0.100.0
uvicorn>=0.30.0
python-multipart>=0.0.6

# LLM integration via OpenRouter (with vision support)
openai==1.55.3
python-dotenv>=1.0.0
httpx==0.27.2

# Utilities
tqdm>=4.60.0
pydantic>=2.0.0

# Optional dependencies (uncomment if needed)
# sentence-transformers>=3.0.0  # For embedding functionality
# anthropic>=0.42.0  # If you want to keep Claude as backup
# dashscope>=1.20.0  # If you want to keep Qwen as backup
